frontend/requirements.txt
pandas==2.2.3
Pillow==11.1.0
Requests==2.32.3
streamlit==1.41.1
